Cleaning Code II
Animation system has a completely new face. I think that we will have
more control that way!

+++ Motor2D/Vector2D.h
@@ -163,7 +163,7 @@ public:
 			angle -= 360;
 		if (angle != 0)
 			angle = -angle;
-		return (float)angle;
+		return angle;
 	}
 
 	Vector2D getNormal() const
